Remove old replaced authentication classes
