+++ launcher/src/main/java/com/skcraft/launcher/auth/AccountList.java
@@ -1,134 +1,78 @@
-/*
- * SK's Minecraft Launcher
- * Copyright (C) 2010-2014 Albert Pham <http://www.sk89q.com> and contributors
- * Please see LICENSE.txt for license information.
- */
-
 package com.skcraft.launcher.auth;
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.beust.jcommander.internal.Lists;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.skcraft.launcher.dialog.component.ListListenerReducer;
 import com.skcraft.launcher.persistence.Scrambled;
 import lombok.Getter;
-import lombok.NonNull;
+import lombok.Setter;
+import lombok.ToString;
+import org.apache.commons.lang.RandomStringUtils;
 
 import javax.swing.*;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Iterator;
+import javax.swing.event.ListDataEvent;
+import javax.swing.event.ListDataListener;
 import java.util.List;
 
 /**
- * A list of accounts that can be stored to disk.
+ * Persisted account list
  */
-@Scrambled("ACCOUNT_LIST")
+@Scrambled("ACCOUNT_LIST_NOT_SECURITY!")
+@Getter
+@Setter
+@ToString
 @JsonIgnoreProperties(ignoreUnknown = true)
-@JsonAutoDetect(
-        getterVisibility = JsonAutoDetect.Visibility.NONE,
-        setterVisibility = JsonAutoDetect.Visibility.NONE,
-        fieldVisibility = JsonAutoDetect.Visibility.NONE)
-public class AccountList extends AbstractListModel<Account> implements ComboBoxModel<Account> {
+public class AccountList implements ListModel<SavedSession> {
+	private List<SavedSession> accounts = Lists.newArrayList();
+	private String clientId = RandomStringUtils.randomAlphanumeric(24);
 
-    @JsonProperty
-    @Getter
-    private List<Account> accounts = new ArrayList<Account>();
-    private transient Account selected;
+	@JsonIgnore private final ListListenerReducer listeners = new ListListenerReducer();
 
-    /**
-     * Add a new account.
-     *
-     * <p>If there is already an existing account with the same ID, then the
-     * new account will not be added.</p>
-     *
-     * @param account the account to add
-     */
-    public synchronized void add(@NonNull Account account) {
-        if (!accounts.contains(account)) {
-            accounts.add(account);
-            Collections.sort(accounts);
-            fireContentsChanged(this, 0, accounts.size());
-        }
-    }
+	public synchronized void add(SavedSession session) {
+		accounts.add(session);
 
-    /**
-     * Remove an account.
-     *
-     * @param account the account
-     */
-    public synchronized void remove(@NonNull Account account) {
-        Iterator<Account> it = accounts.iterator();
-        while (it.hasNext()) {
-            Account other = it.next();
-            if (other.equals(account)) {
-                it.remove();
-                fireContentsChanged(this, 0, accounts.size() + 1);
-                break;
-            }
-        }
-    }
+		int index = accounts.size() - 1;
+		listeners.intervalAdded(new ListDataEvent(this, ListDataEvent.INTERVAL_ADDED, index, index));
+	}
 
-    /**
-     * Set the list of accounts.
-     *
-     * @param accounts the list of accounts
-     */
-    public synchronized void setAccounts(@NonNull List<Account> accounts) {
-        this.accounts = accounts;
-        Collections.sort(accounts);
-    }
+	public synchronized void remove(SavedSession session) {
+		int index = accounts.indexOf(session);
 
-    @Override
-    @JsonIgnore
-    public synchronized int getSize() {
-        return accounts.size();
-    }
+		if (index > -1) {
+			accounts.remove(index);
+			listeners.intervalRemoved(new ListDataEvent(this, ListDataEvent.INTERVAL_REMOVED, index, index));
+		}
+	}
 
-    @Override
-    public synchronized Account getElementAt(int index) {
-        try {
-            return accounts.get(index);
-        } catch (IndexOutOfBoundsException e) {
-            return null;
-        }
-    }
+	public synchronized void update(SavedSession newSavedSession) {
+		int index = accounts.indexOf(newSavedSession);
 
-    @Override
-    public void setSelectedItem(Object item) {
-        if (item == null) {
-            selected = null;
-            return;
-        }
+		if (index > -1) {
+			accounts.set(index, newSavedSession);
+			listeners.contentsChanged(new ListDataEvent(this, ListDataEvent.CONTENTS_CHANGED, index, index));
+		} else {
+			this.add(newSavedSession);
+		}
+	}
 
-        if (item instanceof Account) {
-            this.selected = (Account) item;
-        } else {
-            String id = String.valueOf(item).trim();
-            Account account = new Account(id);
-            for (Account test : accounts) {
-                if (test.equals(account)) {
-                    account = test;
-                    break;
-                }
-            }
-            selected = account;
-        }
+	@Override
+	public int getSize() {
+		return accounts.size();
+	}
 
-        if (selected.getId() == null || selected.getId().isEmpty()) {
-            selected = null;
-        }
-    }
+	@Override
+	public SavedSession getElementAt(int index) {
+		return accounts.get(index);
+	}
 
-    @Override
-    @JsonIgnore
-    public Account getSelectedItem() {
-        return selected;
-    }
+	@Override
+	public void addListDataListener(ListDataListener l) {
+		listeners.addListDataListener(l);
+	}
 
-    public synchronized void forgetPasswords() {
-        for (Account account : accounts) {
-            account.setPassword(null);
-        }
-    }
+	@Override
+	public void removeListDataListener(ListDataListener l) {
+		listeners.removeListDataListener(l);
+	}
 }
